Write mbrng package’s init file on build only
The init file only contains package’s version, which is updated on
build.  Because the init file was tracked with git, when the version
was updated twice, it was unduely marked as dirty.  This patch makes
git to ignore this file and to write it with "unknown" version when
git version cannot be determined and file doesn’t exist already.  It
assumes the init file will never be used for anything else than
package versionning.