Simplify the update.sh script
The update.sh script currently: manages the python environment and dependencies,
updates underlying schema and generates new models, runs tests, commits the changes,
and pushes the changes to the master branch. It is also used to build and publish the
package until the previous commit.

I feel the update.sh script tries to do too much on its own and its releases are
very infrequent, usually one a year or less frequent. I suggest simplifying the
update.sh script to not manage the python environment automatically. Like in SIR,
creating the virtual environment etc. is the responsibility of the developer.
Also, the commits should not be pushed automatically to the master branch.

Further, tools like pipx and uv exist now. We can discuss using those to
automatically manage environments if needed.